+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.1.0
+*2017-05-16* 
+
+- Added functions for creating boostrap cards and tables
+- Navigation entries can be hidden by setting `.visible = False`
+
+
 ## 1.0.1
 
 - Added function `user_has_permission` to acl interface